+++ packages/matter-node.js/src/examples/ComposedDeviceNode.ts
