+++ packages/matter-node.js/src/examples/DeviceNode.ts
@@ -1,219 +1,175 @@
 #!/usr/bin/env node
-
 /**
  * @license
- * Copyright 2022-2023 Project CHIP Authors
+ * Copyright 2022 The node-matter Authors
  * SPDX-License-Identifier: Apache-2.0
  */
 
 /**
- * Important note: This file is part of the legacy matter-node API and should not be used anymore!
- * Please use the new API classes!
+ * This example shows how to create a simple on-off Matter device.
+ * It can be used as CLI script and starting point for your own device node implementation.
  */
 
+/**
+ * Import needed modules from @project-chip/matter-node.js
+ *
+ * When you use this as example please adjust the imports as stated in the "same as ..." comments and simply use
+ * @project-chip/matter-node.js as dependency in your package.json.
+ */
 // Include this first to auto-register Crypto, Network and Time Node.js implementations
-import { MatterDevice } from "../"; // same as @project-chip/matter-node.js
-
+import { CommissionableMatterNode, MatterServer } from "../"; // same as @project-chip/matter-node.js
 import { commandExecutor, getIntParameter, getParameter, requireMinNodeVersion } from "../util"; // same as @project-chip/matter-node.js/util
-import { UdpInterface } from "../net"; // same as @project-chip/matter-node.js/net
+import { Time } from "../time"; // same as @project-chip/matter-node.js/time
+import { OnOffLightDevice, OnOffPluginUnitDevice } from "../exports/device"; // same as @project-chip/matter-node.js/device
+import { VendorId } from "../exports/datatype"; // same as @project-chip/matter-node.js/datatype
 import { Logger } from "../exports/log"; // same as @project-chip/matter-node.js/log
 import { StorageManager, StorageBackendDisk } from "../storage"; // same as @project-chip/matter-node.js/storage
-import { SecureChannelProtocol } from "../exports/securechannel"; // same as @project-chip/matter-node.js/securechannel
-import { PaseServer, CaseServer } from "../exports/session"; // same as @project-chip/matter-node.js/session
-import { ClusterServer, InteractionServer } from "../exports/interaction"; // same as @project-chip/matter-node.js/interaction
-import {
-    BasicInformationCluster, GeneralCommissioningCluster, RegulatoryLocationType, OperationalCredentialsCluster,
-    OnOffCluster, NetworkCommissioningCluster, NetworkCommissioningStatus, AdminCommissioningCluster,
-    CommissioningWindowStatus, AccessControlCluster, GeneralCommissioningClusterHandler,
-    OperationalCredentialsClusterHandler, OnOffClusterHandler, AdminCommissioningHandler, NetworkCommissioningHandler,
-} from "../exports/cluster"; // same as @project-chip/matter-node.js/cluster
-import { VendorId, FabricIndex } from "../exports/datatype"; // same as @project-chip/matter-node.js/datatype
-import { DeviceTypes } from "../exports/device"; // same as @project-chip/matter-node.js/device
-import { MdnsBroadcaster, MdnsScanner } from "../exports/mdns"; // same as @project-chip/matter-node.js/mdns
-import {
-    CommissionningFlowType, DiscoveryCapabilitiesSchema, ManualPairingCodeCodec, QrPairingCodeCodec, QrCode
-} from "../exports/schema";  // same as @project-chip/matter-node.js/schema
-import { AttestationCertificateManager, CertificationDeclarationManager } from "../exports/certificate";  // same as @project-chip/matter-node.js/certificate
-import { Crypto } from "../crypto";  // same as @project-chip/matter-node.js/crypto
-import { Time } from "../time";  // same as @project-chip/matter-node.js/time
 
 const logger = Logger.get("Device");
 
 requireMinNodeVersion(16);
 
-const storage = new StorageBackendDisk(getParameter("store") ?? "device-node");
+const storageLocation = getParameter("store") ?? "device-node";
+const storage = new StorageBackendDisk(storageLocation, getIntParameter("clearstorage") === 1);
+logger.info(`Storage location: ${storageLocation} (Directory)`);
+logger.info('Use the parameter "-store NAME" to specify a different storage location, use -clearstorage to start with an empty storage.')
 
-class DeviceNode {
+class Device {
     async start() {
         logger.info(`node-matter`);
+
+        /**
+         * Initialize the storage system.
+         *
+         * The storage manager is then also used by the Matter server, so this code block in general is required,
+         * but you can choose a different storage backend as long as it implements the required API.
+         */
 
         const storageManager = new StorageManager(storage);
         await storageManager.initialize();
 
+        /**
+         * Collect all needed data
+         *
+         * This block makes sure to collect all needed data from cli or storage. Replace this with where ever your data
+         * come from.
+         *
+         * Note: This example also uses the initialized storage system to store the device parameter data for convenience
+         * and easy reuse. When you also do that be careful to not overlap with Matter-Server own contexts
+         * (so maybe better not ;-)).
+         */
+
         const deviceStorage = storageManager.createContext("Device");
 
+        if (deviceStorage.has("isSocket")) {
+            logger.info("Device type found in storage. -type parameter is ignored.");
+        }
+        const isSocket = deviceStorage.get("isSocket", getParameter("type") === "socket");
         const deviceName = "Matter test device";
-        const deviceType = 257 /* Dimmable bulb */;
         const vendorName = "matter-node.js";
         const passcode = getIntParameter("passcode") ?? deviceStorage.get("passcode", 20202021);
         const discriminator = getIntParameter("discriminator") ?? deviceStorage.get("discriminator", 3840);
-        if (discriminator > 4095) throw new Error("Discriminator value must be less than 4096");
         // product name / id and vendor id should match what is in the device certificate
         const vendorId = new VendorId(getIntParameter("vendorid") ?? deviceStorage.get("vendorid", 0xFFF1));
-        const productName = "matter-node.js Test Product";
+        const productName = `node-matter OnOff ${isSocket ? "Socket" : "Light"}`;
         const productId = getIntParameter("productid") ?? deviceStorage.get("productid", 0x8000);
+
+        const netAnnounceInterface = getParameter("announceinterface");
+        const port = getIntParameter("port") ?? 5540;
 
         deviceStorage.set("passcode", passcode);
         deviceStorage.set("discriminator", discriminator);
         deviceStorage.set("vendorid", vendorId.id);
         deviceStorage.set("productid", productId);
+        deviceStorage.set("isSocket", isSocket);
 
-        // Barebone implementation of the On/Off cluster
-        const onOffClusterServer = ClusterServer(
-            OnOffCluster,
-            { lightingLevelControl: false },
-            { onOff: false }, // Off by default
-            OnOffClusterHandler()
-        );
+        /**
+         * Create Device instance and add needed Listener
+         *
+         * Create an instance of the matter device class you want to use.
+         * This example uses the OnOffLightDevice or OnOffPluginUnitDevice depending on the value of the type  parameter.
+         * To execute the on/off scripts defined as parameters a listener for the onOff attribute is registered via the
+         * device specific API.
+         *
+         * The below logic also adds command handlers for commands of clusters that normally are handled device internally
+         * like identify that can be implemented with the logic when these commands are called.
+         */
 
-        // We listen to the attribute update to trigger an action. This could also have been done in the method invokations in the server.
-        onOffClusterServer.attributes.onOff.addListener(on => commandExecutor(on ? "on" : "off")?.());
+        const onOffDevice = isSocket ? new OnOffPluginUnitDevice() : new OnOffLightDevice();
+        onOffDevice.addOnOffListener(on => commandExecutor(on ? "on" : "off")?.());
 
-        const secureChannelProtocol = new SecureChannelProtocol(
-            await PaseServer.fromPin(passcode, { iterations: 1000, salt: Crypto.getRandomData(32) }),
-            new CaseServer(),
-        );
+        onOffDevice.addCommandHandler("identify", async ({ request: { identifyTime } }) => logger.info(`Identify called for OnOffDevice: ${identifyTime}`));
 
-        const netAnnounceInterface = getParameter("announceinterface");
-        const port = getIntParameter("port") ?? 5540;
+        /**
+         * Create Matter Server and Commissionable Node
+         *
+         * To allow the device to be announced, found, paired and operated we need a MatterServer instance and add a
+         * commissionableNode to it and add the just created device instance to it.
+         * The commissionable node defines the port where the server listens for the UDP packages of the Matter protocol
+         * and initializes deice specific certificates and such.
+         *
+         * The below logic also adds command handlers for commands of clusters that normally are handled internally
+         * like testEventTrigger (General Diagnostic Cluster) that can be implemented with the logic when these commands
+         * are called.
+         */
 
-        const paa = new AttestationCertificateManager(vendorId);
-        const { keyPair: dacKeyPair, dac } = paa.getDACert(productId)
-        const certificationDeclaration = CertificationDeclarationManager.generate(vendorId, productId);
+        const matterServer = new MatterServer(storageManager, netAnnounceInterface);
 
-        const device = new MatterDevice(deviceName, deviceType, vendorId, productId, discriminator, storageManager)
-            .addNetInterface(await UdpInterface.create(port, "udp4"))
-            .addNetInterface(await UdpInterface.create(port, "udp6"))
-            .addScanner(await MdnsScanner.create())
-            .addBroadcaster(await MdnsBroadcaster.create(port, netAnnounceInterface))
-            .addProtocolHandler(secureChannelProtocol)
-            .addProtocolHandler(new InteractionServer(storageManager)
-                .addEndpoint(0x00, DeviceTypes.ROOT, [
-                    ClusterServer(BasicInformationCluster, {}, {
-                        dataModelRevision: 1,
-                        vendorName,
-                        vendorId,
-                        productName,
-                        productId,
-                        nodeLabel: productName,
-                        hardwareVersion: 0,
-                        hardwareVersionString: "0",
-                        location: "US",
-                        localConfigDisabled: false,
-                        softwareVersion: 1,
-                        softwareVersionString: "v1",
-                        capabilityMinima: {
-                            caseSessionsPerFabric: 3,
-                            subscriptionsPerFabric: 3,
-                        },
-                        serialNumber: `node-matter-${Time.nowMs()}`,
-                    }, {}),
-                    ClusterServer(GeneralCommissioningCluster, {}, {
-                        breadcrumb: BigInt(0),
-                        basicCommissioningInfo: {
-                            failSafeExpiryLengthSeconds: 60 /* 1min */,
-                            maxCumulativeFailsafeSeconds: 900 /* Recommended according to Specs */,
-                        },
-                        regulatoryConfig: RegulatoryLocationType.Indoor,
-                        locationCapability: RegulatoryLocationType.IndoorOutdoor,
-                        supportsConcurrentConnections: true,
-                    }, GeneralCommissioningClusterHandler),
-                    ClusterServer(OperationalCredentialsCluster, {}, {
-                        nocs: [],
-                        fabrics: [],
-                        supportedFabrics: 254,
-                        commissionedFabrics: 0,
-                        trustedRootCertificates: [],
-                        currentFabricIndex: FabricIndex.NO_FABRIC,
-                    },
-                        OperationalCredentialsClusterHandler({
-                            devicePrivateKey: dacKeyPair.privateKey,
-                            deviceCertificate: dac,
-                            deviceIntermediateCertificate: paa.getPAICert(),
-                            certificationDeclaration,
-                        }),
-                    ),
-                    ClusterServer(NetworkCommissioningCluster,
-                        {
-                            wifi: false,
-                            thread: false,
-                            ethernet: true,
-                        },
-                        {
-                            maxNetworks: 1,
-                            connectMaxTimeSeconds: 20,
-                            interfaceEnabled: true,
-                            lastConnectErrorValue: 0,
-                            lastNetworkId: Buffer.alloc(32),
-                            lastNetworkingStatus: NetworkCommissioningStatus.Success,
-                            networks: [{ networkId: Buffer.alloc(32), connected: true }],
-                            scanMaxTimeSeconds: 5,
-                        },
-                        NetworkCommissioningHandler(),
-                    ),
-                    ClusterServer(AdminCommissioningCluster,
-                        {
-                            basic: true,
-                        },
-                        {
-                            windowStatus: CommissioningWindowStatus.WindowNotOpen,
-                            adminFabricIndex: null,
-                            adminVendorId: null,
-                        },
-                        AdminCommissioningHandler(secureChannelProtocol),
-                    ),
-                    ClusterServer(AccessControlCluster,
-                        {},
-                        {
-                            acl: [],
-                            extension: [],
-                            subjectsPerAccessControlEntry: 4,
-                            targetsPerAccessControlEntry: 4,
-                            accessControlEntriesPerFabric: 3
-                        },
-                        {},
-                    )
-                ])
-                .addEndpoint(0x01, DeviceTypes.ON_OFF_LIGHT, [onOffClusterServer])
-            );
-        await device.start();
+        const commissionableNode = new CommissionableMatterNode({
+            port,
+            deviceName,
+            deviceType: onOffDevice.deviceType,
+            passcode,
+            discriminator,
+            basicInformation: {
+                vendorName,
+                vendorId,
+                productName,
+                productId,
+                serialNumber: `node-matter-${Time.nowMs()}`,
+            }
+        });
+
+        // optionally add a listener for the testEventTrigger command from the GeneralDiagnostics cluster
+        commissionableNode.addCommandHandler("testEventTrigger", async ({ request: { enableKey, eventTrigger } }) => logger.info(`testEventTrigger called on GeneralDiagnostic cluster: ${enableKey} ${eventTrigger}`));
+
+        commissionableNode.addDevice(onOffDevice);
+
+        matterServer.addCommissionableNode(commissionableNode);
+
+        /**
+         * Start the Matter Server
+         *
+         * After everything was plugged together we can start the server. When not delayed announcement is set for the
+         * commissionable node then this command also starts the announcement of the device into the network.
+         */
+
+        await matterServer.start();
+
+        /**
+         * Print Pairing Information
+         *
+         * If the device is not already commissioned (this info is stored in the storage system) then get and print the
+         * pairing details. This includes the QR code that can be scanned by the Matter app to pair the device.
+         */
 
         logger.info("Listening");
-        if (!device.isCommissioned()) {
-            const qrPairingCode = QrPairingCodeCodec.encode({
-                version: 0,
-                vendorId: vendorId.id,
-                productId,
-                flowType: CommissionningFlowType.Standard,
-                discriminator,
-                passcode,
-                discoveryCapabilities: DiscoveryCapabilitiesSchema.encode({
-                    ble: false,
-                    softAccessPoint: false,
-                    onIpNetwork: true,
-                }),
-            });
-            console.log(QrCode.encode(qrPairingCode));
-            console.log(`QR Code URL: https://project-chip.github.io/connectedhomeip/qrcode.html?data=${qrPairingCode}`);
-            console.log(`Manual pairing code: ${ManualPairingCodeCodec.encode({ discriminator, passcode })}`);
+        if (!commissionableNode.isCommissioned()) {
+            const pairingData = commissionableNode.getPairingCode();
+            const { qrCode, qrPairingCode, manualPairingCode } = pairingData;
+
+            console.log(qrCode);
+            logger.info(`QR Code URL: https://project-chip.github.io/connectedhomeip/qrcode.html?data=${qrPairingCode}`);
+            logger.info(`Manual pairing code: ${manualPairingCode}`);
         } else {
-            console.log("DeviceNode is already commissioned. Waiting for controllers to connect ...");
+            logger.info("Device is already commissioned. Waiting for controllers to connect ...");
         }
     }
 }
 
-new DeviceNode().start().catch(error => logger.error(error));
+new Device().start().then(() => { /* done */ }).catch(err => console.error(err));
 
 process.on("SIGINT", () => {
-    storage.close().then(() => process.exit(0)).catch(() => process.exit(1));
+    // Pragmatic way to make sure the storage is correctly closed before the process ends.
+    storage.close().then(() => process.exit(0)).catch(err => console.error(err));
 });
