+++ packages/matter.js/src/MatterServer.ts
@@ -13,14 +13,30 @@ import { PairableMatterNode } from "./PairableMatterNode.js";
 // TODO Move Mdns instances internally
 // TODO enhance storage manager to support multiple nodes
 
+/**
+ * Main Matter server class that represents the process on the host allowing to commission and pair multiple devices
+ * by reusing MDNS scanner and Broadcaster
+ */
 export class MatterServer {
     private readonly nodes: MatterNode[] = [];
 
+    /**
+     * Create a new Matter server instance
+     *
+     * @param storageManager Storage manager instance to use for all nodes
+     * @param mdnsAnnounceInterface Optional interface to use for MDNS announcements. If not provided announcements will
+     *                              be sent from all network interfaces
+     */
     constructor(
         private storageManager: StorageManager,
         private mdnsAnnounceInterface?: string,
     ) { }
 
+    /**
+     * Add a commissionable node to the server
+     *
+     * @param commisionableNode Commissionable node to add
+     */
     addCommissionableNode(commisionableNode: CommissionableMatterNode) {
         if (this.nodes.length > 0) {
             throw new Error("Only one node is allowed for now");
@@ -40,6 +56,11 @@ export class MatterServer {
         this.nodes.push(commisionableNode);
     }
 
+    /**
+     * Add a pairable node to the server
+     *
+     * @param pairableNode Pairable node to add
+     */
     addPairableNode(pairableNode: PairableMatterNode) {
         if (this.nodes.length > 0) {
             throw new Error("Only one node is allowed for now");
@@ -49,6 +70,10 @@ export class MatterServer {
         this.nodes.push(pairableNode);
     }
 
+    /**
+     * Start the server and all nodes. If the nodes do not have specified a delayed announcement or pairing they will
+     * be announced/paired immediately.
+     */
     async start() {
         // TODO the mdns classes will later be in this class and assigned differently!!
         for (const node of this.nodes) {
@@ -67,6 +92,9 @@ export class MatterServer {
         }
     }
 
+    /**
+     * Close the server and all nodes
+     */
     async close() {
         for (const node of this.nodes) {
             await node.close();
